Remove extraneous SetCurrentDirectory(AppDir) calls.  Their only "benefit" was to cause problems when /y:..\ or some other relative path.
Convert # comments in .ini file to ; and recognize comments starting with ;.  This removes the need for using ;= as a comment introducer, which was actually interpreted as a key of ; and a value of whatever the intended comment was.

Create a "documents" directory under DataDir when SavedOpenDir or SavedSaveDir is either empty, or points to a non-existent directory, and use that directory as the default for the operation.

Cancel the Global Option Check/"Check at start" if /s option appears on command line.

Fix FindFirstFile references to ensure they all do FindClose in all paths.